+++ src/components/InstancesList/InstancesList.test.js
@@ -394,6 +394,33 @@ describe('InstancesList', () => {
 
           expect((screen.getByRole('option', { name: 'Contributors' })).selected).toBeTruthy();
         });
+
+        it('should select Relevance selected sort option when in search query', () => {
+          renderInstancesList({ segment: 'instances' });
+
+          const search = '?segment=instances&sort=relevance';
+          act(() => { history.push({ search }); });
+          openActionMenu();
+
+          const option = within(screen.getByTestId('menu-section-sort-by')).getByRole('option', { name: 'Relevance' });
+          expect(option.selected).toBeTruthy();
+        });
+
+        it('should set aria-sort to none on sorted columns after query sort by Relevance', () => {
+          renderInstancesList({
+            segment: 'instances',
+            parentResources: {
+              ...resources,
+              query: {
+                query: '',
+                sort: 'relevance',
+              }
+            },
+          });
+
+          const sortCols = document.querySelectorAll('[aria-sort="ascending"], [aria-sort="descending"]');
+          expect(sortCols).toHaveLength(0);
+        });
       });
     });